Fix break changes between Select2 v3 and v4
Since its version 5.6.0, ACF has switched to Select2 v4
and thus has introduced some break changes even if we used
its `acf.select2` object.
This is due to the fact that ACF and its object don't abstract
Select2 options by its own and haven't chosen to handle the renaming
of the formatting options `formatResult` and `formatSelection`
(cf.: https://t.co/RhEAwy8Fed?amp=1).

As ACF allows to choose Select2 version (3 or 4) and handles this detail
in its JS process, we should handle this point in JS and CSS as well.